+++ src/main/java/vn/edu/ute/productmgmt/repo/UserAccountRepository.java
@@ -10,5 +10,5 @@ public interface UserAccountRepository {
     List<UserAccount> findAll(EntityManager em) throws Exception;
     void insert(EntityManager em, UserAccount userAccount) throws Exception;
     void update(EntityManager em, UserAccount userAccount) throws Exception;
-    void delete(EntityManager em, int id) throws Exception;
+    void delete(EntityManager em, Long id) throws Exception;
 }
